CHAPTER 7 has been updated.

+++ planner/routes/events.py
@@ -43,12 +43,12 @@ async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
 @event_router.put("/{id}", response_model=Event)
 async def update_event(id: PydanticObjectId, body: EventUpdate, user: str = Depends(authenticate)) -> Event:
     
-    # event = await event_database.get(id)
-    # if event.creator != user:
-    #     raise HTTPException(
-    #     status_code = status.HTTP_400_BAD_REQUEST,
-    #     detail = "Operation not allowred"
-    #     )
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+        status_code = status.HTTP_400_BAD_REQUEST,
+        detail = "Operation not allowred"
+        )
     updated_event = await event_database.update(id, body)
     if not updated_event:
         raise HTTPException(
@@ -59,12 +59,18 @@ async def update_event(id: PydanticObjectId, body: EventUpdate, user: str = Depe
 
 @event_router.delete("/{id}")
 async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
-    event = await event_database.delete(id)
+    event = await event_database.get(id)
     if not event:
             raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Event with supplied ID does not exist"
         )
+    if event.creator != user:
+        raise HTTPException(
+        status_code = status.HTTP_400_BAD_REQUEST,
+        detail = "Operation not allowred"
+        )
+    event = await event_database.delete(id)
     return {
         "message": "Event deleted successfully"
     }
